requirements.txt
absl-py==0.7.0
alabaster==0.7.12
allennlp==0.8.1
apturl==0.5.2
argcomplete==1.8.2
asgiref==3.1.4
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==18.2.0
audiogrep==0.1.5
Automat==0.7.0
aws-xray-sdk==0.95
awscli==1.16.107
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.5.3
bleach==1.5.0
blinker==1.4
boto==2.48.0
boto3==1.9.165
botocore==1.12.165
Bottleneck==1.2.1
bpemb==0.2.12
Brlapi==0.6.7
bs4==0.0.1
cached-property==1.4.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
colorama==0.3.9
command-not-found==0.3
conllu==0.11
constantly==15.1.0
cookies==2.2.1
cryptography==2.7
cssselect==1.0.3
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.10
cytoolz==0.9.0.1
dask==1.1.1
dataclasses==0.6
decorator==4.3.2
defer==1.0.6
defusedxml==0.5.0
dill==0.2.9
distro==1.3.0
distro-info==0.18
Django==2.2.5
django-crispy-forms==1.7.2
djangorestframework==3.10.3
docker==3.7.0
docker-compose==1.21.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
docx==0.2.4
docx2txt==0.6
dryscrape==1.0
EbookLib==0.15
ecdsa==0.13
edit-distance==1.0.3
editdistance==0.5.2
elasticsearch==6.3.1
entrypoints==0.2.3
enum34==1.1.6
fastai==1.0.45
fastprogress==0.1.19
feedfinder2==0.0.4
feedparser==5.2.1
fire==0.1.3
flaky==3.5.3
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
ftfy==5.5.1
future==0.17.1
gast==0.2.2
gensim==3.7.2
gevent==1.3.6
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
hashids==1.2.0
html5lib==0.9999999
httplib2==0.11.3
hyperlink==18.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.10.2
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.4
jsondiff==1.1.1
jsonnet==0.10.0
jsonpickle==1.1
jsonschema==2.6.0
jswipl==1.0.2
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
kafka-python==1.4.6
Keras==2.1.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-vggface==0.5
keyring==15.1.0
keyrings.alt==3.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
louis==3.6.0
lxml==4.2.5
macaroonbakery==1.1.4
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.7.4
mock==2.0.0
more-itertools==6.0.0
moto==1.3.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
mtcnn==0.0.9
murmurhash==1.0.2
mysql-connector-python==8.0.15
mysqlclient==1.3.10
nbconvert==5.2.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.2
newspaper3k==0.2.8
nltk==3.4
nms==0.1.6
nose==1.3.7
notebook==5.0.0
numexpr==2.6.9
numpy==1.17.2
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==2.0.6
olefile==0.46
opencv-python==4.1.0.25
overrides==1.9
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
paramiko==2.6.0
parsel==1.5.1
parsimonious==0.8.0
parso==0.3.4
path.py==10.3.1
pbr==5.1.3
pdf2image==1.6.0
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.1.0
plac==0.9.6
pluggy==0.8.1
pocketsphinx==0.1.3
preshed==2.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.6.3
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.7.0
pyaml==18.11.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.3
pycups==1.9.73
PyDispatcher==2.0.5
pydub==0.23.1
Pygments==2.2.0
PyGObject==3.30.1
PyHamcrest==1.9.0
PyJWT==1.6.4
pymacaroons==0.13.0
pymongo==3.7.2
PyNaCl==1.2.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyRFC3339==1.1
pyswip==0.2.8
pytesseract==0.2.6
pytest==4.3.0
python-apt==1.7.0
python-bidi==0.4.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-debian==0.1.33
python-docx==0.8.10
python-jose==2.0.2
python-pptx==0.6.5
pytorch-pretrained-bert==0.6.2
pytz==2019.2
pyvi==0.0.9.1
PyWavelets==1.0.1
pyxdg==0.25
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.3.1
queuelib==1.5.0
redis==3.3.8
regex==2019.6.8
reportlab==3.5.6
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.1.5
responses==0.10.5
rsa==3.4.2
s3transfer==0.2.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.1
Scrapy==1.6.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.82
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==1.8.3
snap===5.0.0-64-dev-ubuntu18.10-x64-py3.6
snowballstemmer==1.2.1
sortedcontainers==2.1.0
soupsieve==1.9
spacy==2.0.18
SpeechRecognition==3.6.3
spherecluster==0.1.7
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.3.1
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.3
tensorboard==1.12.2
tensorboardX==1.2
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3
textract==1.6.1
texttable==1.4.0
thinc==6.12.1
tika==1.19
tinysegmenter==0.3
tldextract==2.2.1
toolz==0.9.0
torch==1.2.0+cpu
torchvision==0.2.1
tornado==4.5.2
tqdm==4.32.1
traitlets==4.3.2
Twisted==18.9.0
typing==3.6.6
ubuntu-drivers-common==0.0.0
ufw==0.35
ujson==1.35
unattended-upgrades==0.1
Unidecode==1.0.23
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.3
usb-creator==0.3.3
validators==0.12.5
virtualenv==15.1.0
w3lib==1.20.0
wadllib==1.3.3
wcwidth==0.1.7
webencodings==0.5.1
webkit-server==1.0
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.0.2
wikitextparser==0.26.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xkit==0.0.0
xlrd==1.0.0
XlsxWriter==1.1.4
xmltodict==0.12.0
xvfbwrapper==0.2.9
zope.interface==4.6.0
